Add files via upload
This is a review site for games where players can rate online games such as Vigor and Mobile Legends. The users can provide comments regarding their experience and feedback for improvement. It has a list of games, a rating system, and a comment space. Players can view other people's reviews, and features may be optional like user profiles, an admin section, and searching. The program can be designed as a web or mobile application for convenience.